requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
SQLAlchemy
motor
pytest
httpx
python-socketio[asyncio]
python-engineio
websockets
redis
aioredis